daemon: move address `jsonrpc_` methods to an individual module
